Place&UserInDataBaseCommit
