+++ app/src/main/AndroidManifest.xml
@@ -14,14 +14,14 @@
         android:theme="@style/Theme.SmultronStallet"
         tools:targetApi="31">
         <activity
-            android:name=".Database"
+            android:name=".MainActivity"
             android:exported="false">
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name=".DatabaseActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
