--- app/src/main/java/com/example/smultronstallet/Database.kt
@@ -1,11 +1,0 @@
-package com.example.smultronstallet
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class Database : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_database)
-    }
-}
