+++ app/src/main/java/com/example/smultronstallet/User.kt
@@ -1,11 +1,4 @@
 package com.example.smultronstallet
 
 class User(var name: String?,var age: Int?) {
-
-
-
-}
-fun createUser () {
-
-
 }
